orchestration/requirements.txt
apache-airflow==2.5.1
requests==2.28.1
pandas==1.5.3
psycopg2-binary==2.9.5
python-dotenv==0.21.0
